[FIX] account: correct invoice tax totals after changing tax lines
There is the following problem on invoices that are not in company currency.
After changing the tax values in the "Journal Items" tab the tax totals in the "Invoice Lines" tab are not updated accordingly

Reproduce on runbot
  1. Select "BE Company CoA"
  2. New invoice, set currency to USD
  3. Add some item with tax. Remember VAT total
  4. In "Journal Items" edit the tax amount (e.g. put 2000)
  5. Go back to "Invoice Lines". The VAT total line is not updated (Also not after confirming)

Technically the problem is that a account.move.line's amount_currency field is not updated when changing the debit or credit fields.
Thus the data generated from _prepare_tax_lines_data_for_totals_from_invoice contains outdated values.
This leads to a problem in _compute_tax_totals_json:
The tax_lines_data argument that is passed to _get_tax_totals contains outdated values.
Thus the tax totals are computed incorrectly there.

task-3601300