[ADD] pos_cache: a new POS cache module
This new module comes from a need for better performances
on databases with a high amount of data.
Even by loading things in chunk, the POS will struggle to
get everything loaded and the performances will be slow.

This new cache system mainly aims at pre-processing the
computations during the low hours of the day using a cron,
and then serve these precomputed data directly to the POS
client.

This brings improvements on the loading time, as well as
reduce heavily the load on the server when opening POS.

Task id #3483954