[FIX] web_editor: improve edge scrolling in snippets editor
Since [1], the scrolling in the snippets editor is slow. This is because
the drag and drop uses the default values of edge scrolling which are
mostly appropriate for dragging in the backend, where the scroll
distance is usually small.

This commit increase the "scrolling zone" and the "scrolling speed" for
the snippets menu and snippets editor.

[1]: https://github.com/odoo/odoo/commit/7594d71ca8610d5947e80f325ccb57abc23c2c76

task-3607685