[FIX] web_editor: parseHTML DOMParser constructor
Commit [1] did not do the job completely: while the document for
creating the document fragment was the one received as parameter, the
DOMParser constructor remained the one from the window in which the JS
code is executed, which might be different from the received document's
window.  This mismatch happens in particularly when the editable is
inside an iframe.

[1]: https://github.com/odoo/odoo/commit/c74d01279e49fadc5f509cfc3f6baf43d2b532d1