[IMP] account: make currency_rate of account.move.line stored
This i.e. enables to the of this field in SQL queries.

Thus it can i.e. be used to efficiently compute the total of the line
including taxes in company currency (from the total in move currency).
This is needed to display that value in the account invoice report
(see other commits of this PR / task)

task-3613358