[FIX] account: fix single bill creation with PDF and XML from alias
Steps to reproduce:
- Install Accounting
- Configure an Email alias for Vendor Bills journal
- Send an email to the alias with an PDF containing an embedded EDI XML and
the EDI XML corresponding to the PDF as attachments
=> 2 bills are created (one with the PDF as attachment and one with the XML),
instead of one with both PDF and XML as attachments

Cause:
During the processing of the 2 attachments to determine in which bill they
will be linked to, a XML is retrieved from the PDF while decoding it.
This XML is handled like the 2 attachments and messes up with the process.

Solution:
Ignore the XML embedded in the PDF.

opw-3604697