[IMP] account_edi_ubl_cii: add tax prediction if available
The aim of this commit is to improve the quality of the tax retrieving
of the edi import.

Context:
The tax prediction happens when the invoice is given manually or with
the OCR but doesn't happen when it's imported by EDI because the tax is
already set by the OCR.

Before this commit:
During EDI import, the tax retrieved could always be the same even if
the user make some modification to the previous bill.

After this commit:
The tax retrieved leverage the previous data to get better quality
result for tax retrievement.

task-id: 3531473