[FIX] website: edit HTML in default website language
Since the introduction of delay translation in [1] only the edited
master/default language version of the page was being stored in the
database's JSONB `arch_db`.

Because of this if the default language of a website was not `en_US`,
after edit the `en_US` version of the page remained the empty page.
But since [2] the HTML editor edits (reads and writes) the page in
`en_US` only, it was therefore editing the empty page instead of the
actual "master" one. This led to destroying the actual content that
existed in other languages when saving such a modified page from the
HTML editor.

To avoid this [3] temporarily disabled the delay translation feature.

With the delay translation re-activated by the previous commit, it now
makes sense to edit the HTML using the default website language.

Steps to reproduce:
- Add Dutch as the default language of a website.
- Create a new Dutch page.
- Drop an accordion snippet.
- Save.
- Edit the page using the HTML editor. => Empty page is being edited.
- Add a blank line.
- Save.

=> Content was lost.

[1]: https://github.com/odoo/odoo/commit/2d08f97c0778469b409fca23f2be5f5a98ce3df8
[2]: https://github.com/odoo/odoo/commit/fdb9f8273be62d0a6d8051f7ca0a66cdf22ae5e7
[3]: https://github.com/odoo/odoo/commit/0e0a74f8c5fc9f45311e629a76608c6c986d635d

opw-3493355
task-3621753