[IMP] core: warn about compute methods mixing readonly and non-readonly fields
We explicitly discourage a compute method used for both readonly and
non-readonly fields, because it makes it hard to understand whether the
fields should be recomputed when writing on the field's dependencies.

The principle is: when writing on a computed field and its dependency,
the field should be recomputed if and only if it is readonly.  Now
consider two fields A (readonly) and B (non-readonly) computed by the
same method.  Those fields should be recomputed when writing on one of
their dependencies but not on B, because writing on B prevents the
recomputation of both fields.  This awkward interdependency between
fields A and B makes the recomputation mechanism very difficult to
grasp.

So, in order to remote this unnecessary complexity, we advocate that the
fields computed by a given method should either be all readonly, or
all non-readonly.