[IMP] {im, website}_livechat, *: rely on server instead of local state
*: mail.

The live chat embed stores the thread in a cookie and restores it when
the page is reloaded. This is really bad:
- A lot of data are stored as a cookie and add unnecessary overhead to
requests.
- Local state is unreliable.
- Local state is hard to maintain.
- Local state enforces differences between live chat and discuss which
leads to highly specific code.
- Local state can be outdated when the Odoo version changes.

Now that the live chat uses discuss guest system, we can rely on the
result of `init_messaging` to retrieve the channel infos.

From now on, only the `id` and the `model` of the live chat thread will
be stored.

In order to rely as much as possible on the server:
- some formatters were adapted to provide every required data such as
`id/model/fold_state` for temporary threads.
- some methods were adapted to support guests such as `_channel_fold`.

This work will allow future improvements such as better cross-tab
synchronization (currently perfunctory).

task-3488100