[FIX] point_of_sale: show relevant entries in followup
In PR #144945 entries from misc. journals were excluded from the followup.
This exclusion of misc. journal entries broke the followup for
"Customer Account" sales in PoS.
After this commit the followup functionality is restored.

The problem:
PoS uses a misc. journal to track uninvoiced sales / payments.
These misc. entries are i.e. used for the "Customer Account"
payment method (unpaid / due later). It needs to be tracked and
can / should be followed up on. This also applies to the (possible)
reversal of a customer account payment entry. Such a reversal is
generated in case an actual invoice is created.

The fix:
This commit extends the compute function of the field on account move
lines that determines whether the lines is excluded or not.
Any payable / receivable line that would be blocked by the original compute
function but belongs to a PoS journal is unblocked.

task-3632004