[IMP] point_of_sale,pos_*: split components due to perfomance issues
*:l10n_fr_pos_cert,pos_loyalty,pos_online_payment,pos_restaurant,
pos_sale

The product_screen component was too big and was causing performance
issues when UI was rerender. This commit splits it into 3 components:
- product_list
- product_card_wrapper
- order_summary

Statistics with a database of 6,000 products loaded into PoS:

When a product was added to the current order, the `product_screen`
component took ~150ms to render.

With these modifications it now only takes ~3ms.