[FIX] mail, im_livechat: do not overwrite tab title
Since [1], messages received out of focus are notified to the user via
the tab title. In order to do so, the title service is used. This
service changes the tab title according to its internal state.

When the live chat is embed on a page, the original title of the page
is not coming from the title service. When the service state changes,
the title is replaced without taking into account the original title.
This behavior is not correct.

In practice, this functionnality is only used in by the `web` bundle.
The error occurs because the code handling this behavior is placed in
the common bundle while it should not. This PR fixes this issue by
moving this code to the `web` bundle.

task-3644431

[1]: https://github.com/odoo/odoo/commit/f818e50527e38fb831cbe76a4710c