[FIX] tx: add entry in tx file
When adding a new (not l10n module) into stable that has terms that should be
translated. An entry need to be added to the .tx/config file or the new modules
will not be added to Transifex.

During this pr (https://github.com/odoo/odoo/pull/143758), the entry was missing
so this pr will add it.

task: no task