[IMP] core: update noupdate flag when switched to True
When a record is originally in a noupdate=0 block (the default) but we
make it part of a noupdate=1 block in a newer version of a module then
we should update the noupdate flag in ir_model_data. In effect this flag
may be used by other parts of the system (e.g. during upgrades) to
understand whether a record can be safely updated or not.

Co-authored-by: "Christophe Simonis" <chs@odoo.com>