[FIX]purchase_stock: Changed decorator in product_replenish to prevent database looping.
Changed the way the warning for the lack of vendors in a product is chosen to be displayed, so the page doesn't get
stuck looping through an empty list.

Currently both in saas-16.4 and 17.0 in product_template, when the only route available for a product replenishment is "Buy" and there are no vendors assigned to the product, when pressing the "Replenish" button, the page gets stuck trying to display the wizard, but never ends loading it.

Steps to reproduce:

    1. Go to stock or sale modules, to the products section.
    2. Select "Buy" route.
    3. Leave "Purchase" tab empty.
    4. Click on the "Replenish" button.

Description of the issue/feature this PR addresses:
Empty list of vendors prevents the replenishment wizard from being displayed.

Current behavior before PR:
Buttons in status bar and header get blocked and the wizard is never displayed. Only exiting the view and going back into it unlocks the buttons, but doesn't solves the problem.

Desired behavior after PR is merged:
The wizard is displayed as intended.

opw-3653714