[FIX] mail: fix non deterministic im status test
Before this PR the "Can handle im_status of unknown partner" test was
sometimes failing. This test checks that the persona is created when
it's im_status is received if it was not already known. In order to
achieve this, a notification is sent and the `nextTick` helper is used
to wait for the persona to be asserted. This is weak since there is no
way to guarantee a tick is enough to receive and process the
notification.

This PR fixes this issue by replacing the `nextTick` helper by the
`waitNotifications` one that resolves once the notification is
received.

fixes runbot-52628