[FIX] base: Do not use the domains of fields from a related
Issue: Domains are imported for related fields in views that do not
correspond to them. The environment is not that provided for in the field
declaration if the model is different.

E.g.
Model1:
A = fields.Char()
B = fields.Many2one('Model0', domain="[('field', '=', A)]")

Model2
C = fields.Many2one('Model1')
D = fields.Many2one(related="C.B")