[CLN] sale_team: remove the team_id from the res_partner view
Purpose
=======
Clean up the Sales Team/Partner relationship following some feedback that proves
it's currently a little messy.

Specifications 
===============
1) Remove the team_id field from the res.partner model
Rationale: It is probably due to a previous implementation but from our tests
and research, this field has never used

2) Remove the field from the views (and there seem to be quite a few messy ones)
Rationale: It's never used and is just confusing, users expect this team to be
selected on all leads, quotes, etc that are created but that's not the case.
=> We use the user_id field in some cases, but never the team_id (we take the
team of the salesperson or the first one in the list otherwise)

Upgrade PR: https://github.com/odoo/upgrade/pull/5609

Task-3618859