[FIX] l10n_it_edi: fix move_type when importing outgoing invoices
Currently we assume that all imported moves are incoming moves.
This was i.e. done since the tax agency only sends users bills.

But some clients import invoices from other software (i.e. onboarding/starting).

After this PR we check whether the company issuing the imported
invoice is the active company or not.
Depending on this we decide whether the invoice is outgoing or ingoing.

The company issuing the imported invoice is determined by looking at
the VAT number (IdCodice) and Codice Fiscale (CodiceFiscale) stated in the XML.

task-3650355