[IMP] spreadsheet_account: Add formula ODOO.RESIDUAL
Problem
---------
There is currently no formula to get the residual amount of accounts in
a given period of time in spreadsheets.

Objective
---------
Add the formula ODOO.RESIDUAL that does just that.

It should take the following paramters:
- account_codes (mandatory)(can be more than one separate by ,): could
be optional, in that case it could set the domain to all
receivables/payables accounts
- due_date_range or expected_date (optional): date range referred to due
dates (not AML date)
- offset (optional): adding the possibility of the offset to be not only
“Year” but also, a specific date or other periods like Month, Quarter.
- company_id (optional)
- include_unposted (optional) (default: false)

Solution
---------
1. Define the spreadsheet function with the correct parameters in
`accounting_function.js`
2. Hook it to the data source so that it can perform server calls
3. Craft that server python function to returns the list a residual
amounts for the given list of parameters.
4. Adds some tests.

Task-3679808