[FIX] resource: prevent traceback if  number of days is zero
This traceback arises when a user tries to add a line in the 'Working Hours'

Steps to produce

1. Install hr_attedence
2. Go to 'Settings/Technical/Resource/Working Schedules'
3. Click on 'New' and switch to 'Two Weeks Calendar'
4. Delete all records in Working Hours and add a new line

Error:
 ZeroDivisionError
float division by zero

When the user tries to add a new line in Working Hours, a traceback will
be generated because the number_of_days is zero before performing the division.

which leads to traceback from here
https://github.com/odoo/odoo/blob/851eb198433d4fd6f72100568be3bf57a4803064/addons/resource/models/resource_calendar.py#L185

After applying this commit, avoid the division by zero error
because you are dividing by at least 1.

sentry-4889879167