[FIX] mail: chatter using wrong classes and unnecessary style
While working on the chatter spacing task I noticed unnecessary class
on the activitylist component. `fe-3` is not an utility class and 
changing the font-size on the element hr did not have any visual effect.

Moreover the SuggestRecipientsList component was using unnecessary
style since the same spacing can be done with a `ms-5` utility class.

task-