[FIX] web: compare months with month's index
Before this commit, if you compare 2 years and don't have data
for all the months, the comparison will be offset.
Example: you have data in 2023 from January to November and 2022
from March to December, then the graph's comparison will start with
January 2023 and compare it with the first entry of 2022, so March.

After this commit, the offset is fixed. When there is a comparison
by month, it will use the month's index instead of the index of the
data in the results.

Task-ID: 3425017