[IMP] stock,product_expiry: Add barcode separator
This commit adds a new system parameter: `stock.barcode_separator`
Its default value is `,`.

Its purpose is to print serial and lot numbers contained in a package in
a single data matrix by separate each barcode string by the given
character.

The report `stock.report_package_barcode` was adapted to use this
parameter and print a composed barcode (as a DataMatrix) at the end of
each product's lines.
This DataMatrix will contain the product barcode, the quantity and all
the serial/lot numbers.

task-3595212