[FIX] web_editor: toolbar flick on hover reset button
Before this commit, the toolbar could ficker when hovering over the
colorpicker's "reset" button, regardless of the selection direction, if
the selection still contained font elements after the call to the
"applyColor" commmand with a falsy `color` parameter. This can happen,
for example, when hovering over the Background Color colorpicker's
"reset" button while having a selection containing a font element for
foreground color.  This would make the selection to be set around the
remaining font element, leading to a toolbar repositioning.

task-3700731