[FIX] web_editor: toolbar flicker on selection width change
The selection's client rectangle width of a width can change slightly
when hovering over the "reset" button on a selection containing, for
example, a font element with "font-weight: bolder" as style, leading the
toolbar to flicker. For this reason, this commit introduces a condition
in `_positionToolbar` to avoid updating the toolbar's horizontal
position for small changes when the colorpicker is opened.

task-3700731