[IMP] tools: partial backport of cache refactoring
Following multiple ormcache, some cleanup can be made in the ormcache
management, mainly to limit memory usage keeping useless context.

Also, part of the api evolve in a organic way when some cleanup could be
done in the current version.

Backporting stable safe part of this change in 16.4 to keep a similar
structure for multiple ormcache and ease future forwardport.

Still discussing if the stats should be stored on the cache itself
instead of the global STAT, old school coding.