[IMP] base: Remove `user_has_groups` and the fallback in `has_group`
Replaced by:

`_has_group` method on res.users

Before this commit:

The user_has_groups method calls `self.env.user.has_group`, with the
difference being a split (and the negative group '!' used only for the
views) of the provided string and management of `group_no_one` using
`request.session.debug`. The use of `user_has_group` is quite
confusing because it does not relate to the current record but to the
environment.

The `has_group` method is also confusing because if it is applied on an
empty recordset then there is a fallback on `env.user` otherwise it uses
the current record.

The change makes it possible to unify the API and remove ambiguity and
make method calls explicit, and make the groups independent of the
session.

Upcoming changes being the removal of group_no_one and decreasing the
number of session.debug usages.