[FIX] resource, *: add a _get_calendar method
*:
- hr_contract
- hr_holidays

Before this commit, accessing the employee's
calendar was done through employee.resource_calendar_id
However, this might be erroneous when hr_contract is
installed, as the calendar of the employee might be
different given the date for which we want to get the calendar.

This commit implements in `resource.mixin` a new method
called `_get_calendar` taking a date as argument and returns
a recordset of calendar corresponding to one calendar
per employee.