[FIX] sale, purchase, product: show sale warning in catalog
Currently, when adding a product to a quotation with the catalog, no warning message will appear when adding a poduct that has a warning or blocking message.

Steps to reproduce:
-------------------
* Go to **Sales** app -> Configuration -> Settings
* Enable **Sale Warnings**
* Go to **Products** -> Products
* Create a new product
* Under **Sales** tab:
  * For warning, select either warning or blocking message
  * Write a message
* Go to **Orders** -> Quotations
* Create a new quotation
* Select the **Catalog** to add products
* Add the newly created product

Why the fix:
------------
The first version of this fix was a python exclusive change. It was a bit hacky; it was raising an error when the product had a blocking warning and sending a message on the bus when the product had a non-blocking warning.

The second version was changing the return value of `_update_order_line_info` to return the price and the warning if any. The warning was shown inside JS with `_updateQuantity`. This change was not good for a stable verion as it was changing the signature of a public method.

This fix automatically changes the data that is loaded to the catalog. It adds the waning message if any and changes the `readOnly` field accordingly.

For products with blocking warnings, the data loaded to the catalog will mark the product as read only and display the warning message directly in the catalog, on the product. For products with non-blocking warnings, the warning message will be loaded when updating the quantity.

Regarding the field `readOnly`:
```python
        for product_id, price in pricelist.items():
            product = self.env['product.product'].browse(product_id)
            res[product_id] = {
                'price': price,
                'warning': product.sale_line_warn_msg,
            }
            if product.sale_line_warn == "block":
                res[product_id]['readOnly'] = True
        return res
```
```python
        for product_id, price in pricelist.items():
            product = self.env['product.product'].browse(product_id)
            res[product_id] = {
                'price': price,
                'warning': product.sale_line_warn_msg,
                'readOnly': product.sale_line_warn == "block"
            }
        return res
```
The first solution is prefered over the second as it will not override the default value of the field. Default data gets overrinden by the new data here:

https://github.com/odoo/odoo/blob/21c25a7ccd0ba2d6574ddbcfbcf50dbbc03a1e6c/addons/product/models/product_catalog_mixin.py#L97-L99

For example, in the case where the sale order is cancelled or locked, the field will remain `True`.

https://github.com/odoo/odoo/blob/21c25a7ccd0ba2d6574ddbcfbcf50dbbc03a1e6c/addons/sale/models/sale_order.py#L1883-L1893

Since the module purchase also uses the warning on products, this fix is extended to include it.

opw-3631511