[REF] stock,product_expiry: remove removal strategy model
It's not needed to have an obejct for that. We usualy adds
new object when we need to hide/show parts of the selection
in different cases. However for removal strategy we don't use
this feature.

On top of that. It causes issues since the method field (aka
the technical name) is translatable.