[FIX] web: resolve grid view time format for minute
Versions:
-master

Steps to reproduce:
-Go to the timesheet grid view
-Click on grid cell and put time

Issue:
Users could input any number of digits for minutes, and that was converted to
hours.

Cause:
Lack of validation for the number of digits in minutes during time parsing.

Solution:
Added a condition to restrict users from entering more than two digits for
minutes.

task-3429403