[FIX] hr_timesheets: fixes display name recompute issue
Steps to reproduce:
-------------

- install project and hr_timesheet
- project -> Configuration -> Setting -> Allow timesheets in Time Management
- open all tasks/my tasks  in kanban view
- select on Quick Add Task
- add task name , then select any project
- when changing the project, the task name field automatically becomes blank.

Issue:
-------------
the task name goes empty after selecting a project, even if you rewrite it and
switch projects. This only happens in the Quick Add feature in the kanban
view after enabling timesheets in the  project settings.

Cause:
-------------
The allow_timesheets field's calculation depends on project_id.allow_timesheets,
and this field also utilized in  depends of the _compute_display_name method.
When the project changes,this triggers a recompute, resulting in the task name
becoming blank.

Solution :
----------------
remove 'allow_timesheets' in the depends of the _compute_display_name method.

task-3706741